+++ examples/games/moba/character-base.js
@@ -144,7 +144,7 @@ export class Character extends Hero {
           z: startUpPositions[this.core.player.data.team][2]
         },
         name: this.name,
-        texturesPaths: ['./res/meshes/glb/textures/mutant_origin.webp'],
+        texturesPaths: ['res/textures/default.png'],
         raycast: {enabled: true, radius: 5},
         pointerEffect: {
           enabled: true,